// apps/studio/lib/api/self-hosted/auth-config/fields.ts
export type GoTrueConfigFieldType = 'string' | 'number' | 'boolean'
// ...
export const GOTRUE_DURATION_FIELDS: Record<string, 'seconds' | 'hours' | 'per_hour'> = {
  API_MAX_REQUEST_DURATION: 'seconds',
  SESSIONS_INACTIVITY_TIMEOUT: 'hours',
  SESSIONS_TIMEBOX: 'hours',
  MFA_PHONE_MAX_FREQUENCY: 'per_hour',
  SMS_MAX_FREQUENCY: 'per_hour',
  SMTP_MAX_FREQUENCY: 'per_hour',
}